add SceneManager
