+++ app/src/main/java/com/example/testsensors/GamePanel.java
@@ -13,35 +13,18 @@ import android.view.SurfaceView;
 public class GamePanel extends SurfaceView implements SurfaceHolder.Callback {
 
     private MainThread thread;
-    private Rect r = new Rect();
 
-    private RectPlayer player;
-    private Point playerPoint;
-    private ObstacleManager obstacleManager;
-
-    private boolean movingPlayer = false;
-    private boolean gameOver = false;
-    private long gameOverTime;
+    private SceneManager sceneManager;
 
     public GamePanel(Context context) {
         super(context);
         getHolder().addCallback(this);
 
         thread = new MainThread(getHolder(), this);
-        player = new RectPlayer(new Rect(100, 100, 200, 200), Color.rgb(255, 0, 0));
-        playerPoint = new Point(Constants.SCREEN_WIDTH/2, 3 * Constants.SCREEN_HEIGHT / 4);
-        player.update(playerPoint);
 
-        obstacleManager = new ObstacleManager(200, 350, 75, Color.BLACK);
+        sceneManager = new SceneManager();
 
         setFocusable(true);
-    }
-
-    public void reset() {
-        playerPoint = new Point(Constants.SCREEN_WIDTH/2, 3 * Constants.SCREEN_HEIGHT / 4);
-        player.update(playerPoint);
-        obstacleManager = new ObstacleManager(200, 350, 75, Color.BLACK);
-        movingPlayer = false;
     }
 
     @Override
@@ -73,69 +56,19 @@ public class GamePanel extends SurfaceView implements SurfaceHolder.Callback {
 
     @Override
     public boolean onTouchEvent(MotionEvent event) {
-        switch (event.getAction()) {
-            case MotionEvent.ACTION_DOWN:
-                if (!gameOver && player.getRectangle().contains((int) event.getX(), (int) event.getY())) {
-                    movingPlayer = true;
-                }
-                if (gameOver && System.currentTimeMillis() - gameOverTime >= 2_000) {
-                    reset();
-                    gameOver = false;
-                }
-                break;
-            case MotionEvent.ACTION_MOVE:
-                if (!gameOver && movingPlayer) {
-                    playerPoint.set((int) event.getX(), (int) event.getY());
-                }
-                break;
-
-            case MotionEvent.ACTION_UP:
-                movingPlayer = false;
-                break;
-        }
+        sceneManager.recieveTouch(event);
 
         return true;
         //return super.onTouchEvent(event);
     }
 
     public void update() {
-        if (!gameOver) {
-            player.update(playerPoint);
-            obstacleManager.update();
-
-            if (obstacleManager.playerCollide(player)) {
-                gameOver = true;
-                gameOverTime = System.currentTimeMillis();
-            }
-        }
-
+        sceneManager.update();
     }
 
     @Override
     public void draw(Canvas canvas) {
         super.draw(canvas);
-
-        canvas.drawColor(Color.WHITE);
-
-        player.draw(canvas);
-        obstacleManager.draw(canvas);
-
-        if (gameOver) {
-            Paint paint = new Paint();
-            paint.setTextSize(100);
-            paint.setColor(Color.CYAN);
-            drawCenterText(canvas, paint, "GAME OVER");
-        }
-    }
-
-    private void drawCenterText(Canvas canvas, Paint paint, String text) {
-        paint.setTextAlign(Paint.Align.LEFT);
-        canvas.getClipBounds(r);
-        int cHeight = r.height();
-        int cWidth = r.width();
-        paint.getTextBounds(text, 0, text.length(), r);
-        float x = cWidth / 2f - r.width() / 2f - r.left;
-        float y = cHeight / 2f + r.height() / 2f - r.bottom;
-        canvas.drawText(text, x, y, paint);
+        sceneManager.draw(canvas);
     }
 }
