+++ app/src/main/java/com/example/testsensors/Scene.java
@@ -1,0 +1,11 @@
+package com.example.testsensors;
+
+import android.graphics.Canvas;
+import android.view.MotionEvent;
+
+public interface Scene {
+    public void update();
+    public void draw(Canvas canvas);
+    public void terminate();
+    public void revieveTouch(MotionEvent event);
+}
